requirements.txt
Flask~=3.1.2
twilio~=9.8.0
python-dotenv~=1.1.1
gunicorn
flask_mail